pi_requirements.txt
Flask>=3.0
python-escpos>=3.0
Pillow>=10.0
python-dotenv>=1.0

